# src/kst-platform/disease-warning/src/main/java/de/fhg/iese/kickstarttrustee/diseasewarning/api/dto/CreateDiseaseWarningSettingsDTO.java
package de.fhg.iese.kickstarttrustee.diseasewarning.api.dto;

public record CreateDiseaseWarningSettingsDTO(boolean isDiseaseReportProcessingEnabled,
        boolean isDiseaseWarningCreationEnabled, boolean isEmailNotificationEnabled) {
}