# src/kst-platform/disease-warning/src/main/java/de/fhg/iese/kickstarttrustee/diseasewarning/api/dto/UpdateDiseaseWarningSettingsDTO.java
package de.fhg.iese.kickstarttrustee.diseasewarning.api.dto;

public record UpdateDiseaseWarningSettingsDTO(boolean isDiseaseReportProcessingEnabled,
        boolean isDiseaseWarningCreationEnabled, boolean isEmailNotificationEnabled) {
}