# src/kst-platform/ui/owner/owner-api/api-docs.json
{"openapi":"3.0.1","info":{"title":"Kickstart Trustee","description":"This is the API of the Kickstart Trustee backend","version":"v1"},"servers":[{"url":"http://localhost:8090","description":"Generated server url"}],"security":[{"security_auth":[]}],"paths":{"/service/disease-warning/settings/{id}":{"get":{"tags":["service.disease-warning"],"operationId":"getSettingsById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiseaseWarningSettingsDTO"}}}}}},"put":{"tags":["service.disease-warning"],"operationId":"updateSettings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateDiseaseWarningSettingsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiseaseWarningSettingsDTO"}}}}}}},"/owner/profile/{id}":{"get":{"tags":["owner"],"operationId":"getProfileById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerProfileDTO"}}}}}},"put":{"tags":["owner"],"operationId":"updateMyProfile","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOwnerProfileDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerProfileDTO"}}}}}}},"/service/disease-warning/settings":{"post":{"tags":["service.disease-warning"],"operationId":"createOwnSettings","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDiseaseWarningSettingsDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiseaseWarningSettingsDTO"}}}}}}},"/provider/data":{"post":{"tags":["provider.data"],"operationId":"createDataItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDataItemDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}},"/provider/data/bulk":{"post":{"tags":["provider.data"],"operationId":"createDataItems","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateDataItemDTO"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}}},"/provider/consent-request":{"get":{"tags":["provider.consent"],"operationId":"getAllRequestedConsentRequests","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProviderConsentRequestDTO"}}}}}}},"post":{"tags":["provider.consent"],"operationId":"createConsentRequest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProviderCreateConsentRequestDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProviderConsentRequestDTO"}}}}}}},"/prosumer/data":{"get":{"tags":["prosumer.data"],"operationId":"getDataItems","parameters":[{"name":"dataItemType","in":"query","required":true,"schema":{"type":"string"}},{"name":"ownerId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}},"post":{"tags":["prosumer.data"],"operationId":"createDataItem_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDataItemDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}},"/prosumer/data/bulk":{"post":{"tags":["prosumer.data"],"operationId":"createDataItems_1","requestBody":{"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateDataItemDTO"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}}},"/prosumer/consent-request":{"get":{"tags":["prosumer.consent"],"operationId":"getAllRequestedConsentRequests_1","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProsumerConsentRequestDTO"}}}}}}},"post":{"tags":["prosumer.consent"],"description":"Additional rules for the consent request creation:\n\n- If you create a consent request with a not empty consumedDataItemTypes field, then you must provide a dataUsageStatement and a purpose.\n\n- If you create a consent request and an empty consumedDataItemTypes field (this means you requested for data provision only), then you can omit the dataUsageStatement and the purpose.\n","operationId":"createConsentRequest_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProsumerCreateConsentRequestDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProsumerConsentRequestDTO"}}}}}}},"/owner/profile":{"post":{"tags":["owner"],"operationId":"createMyProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOwnerProfileDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerProfileDTO"}}}}}}},"/consumer/consent-request":{"get":{"tags":["consumer.consent"],"operationId":"getAllRequestedConsentRequests_2","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConsumerConsentRequestDTO"}}}}}}},"post":{"tags":["consumer.consent"],"operationId":"createConsentRequest_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsumerCreateConsentRequestDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConsumerConsentRequestDTO"}}}}}}},"/provider/consent-request/{id}":{"get":{"tags":["provider.consent"],"operationId":"getRequestedConsentRequestById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProviderConsentRequestDTO"}}}}}},"patch":{"tags":["provider.consent"],"operationId":"updateRequestedConsentRequest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConsentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/prosumer/consent-request/{id}":{"get":{"tags":["prosumer.consent"],"operationId":"getRequestedConsentRequestById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProsumerConsentRequestDTO"}}}}}},"patch":{"tags":["prosumer.consent"],"operationId":"updateRequestedConsentRequest_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConsentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/owner/consent/{id}":{"get":{"tags":["owner.consent"],"operationId":"getOwnedConsentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerConsentDTO"}}}}}},"patch":{"tags":["owner.consent"],"operationId":"updateOwnedConsent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConsentDTO"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/owner/consent-request/{id}":{"get":{"tags":["owner.consent"],"operationId":"getOwnedConsentRequestById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerConsentRequestDTO"}}}}}},"patch":{"tags":["owner.consent"],"operationId":"updateOwnedConsentRequest","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConsentRequestDTO"}}},"required":true},"responses":{"204":{"description":"No Content"}}}},"/consumer/consent-request/{id}":{"get":{"tags":["consumer.consent"],"operationId":"getRequestedConsentRequestById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConsumerConsentRequestDTO"}}}}}},"patch":{"tags":["consumer.consent"],"operationId":"updateRequestedConsentRequest_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateConsentRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/service/disease-warning/settings/own":{"get":{"tags":["service.disease-warning"],"operationId":"getOwnSettings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiseaseWarningSettingsDTO"}}}}}}},"/provider/consent":{"get":{"tags":["provider.consent"],"operationId":"getAllConsents","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","REVOKED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProviderConsentDTO"}}}}}}}},"/provider/consent/{id}":{"get":{"tags":["provider.consent"],"operationId":"getConsentById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProviderConsentDTO"}}}}}}},"/prosumer/data/{id}":{"get":{"tags":["prosumer.data"],"operationId":"getDataItemById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}},"/prosumer/data/{id}/meta":{"get":{"tags":["prosumer.data"],"operationId":"getMetaDataById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MetaDataDTO"}}}}}}},"/prosumer/consent":{"get":{"tags":["prosumer.consent"],"operationId":"getAllConsents_1","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","REVOKED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProsumerConsentDTO"}}}}}}}},"/prosumer/consent/{id}":{"get":{"tags":["prosumer.consent"],"operationId":"getConsentById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProsumerConsentDTO"}}}}}}},"/owner/profile/me":{"get":{"tags":["owner"],"operationId":"getMyProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerProfileDTO"}}}}}}},"/owner/data":{"get":{"tags":["owner.data"],"operationId":"getDataItems_1","parameters":[{"name":"dataItemType","in":"query","required":false,"schema":{"type":"string"}},{"name":"providerId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}}},"/owner/data/{id}":{"get":{"tags":["owner.data"],"operationId":"getDataItemById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataItemDTO"}}}}}},"delete":{"tags":["owner.data"],"operationId":"deleteDataItemById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}},"/owner/data/{id}/meta":{"get":{"tags":["owner.data"],"operationId":"getMetaDataById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MetaDataDTO"}}}}}}},"/owner/consent":{"get":{"tags":["owner.consent"],"operationId":"getOwnedConsents","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","REVOKED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerConsentDTO"}}}}}}}},"/owner/consent-request":{"get":{"tags":["owner.consent"],"operationId":"getOwnedConsentRequests","parameters":[{"name":"requesterId","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerConsentRequestDTO"}}}}}}}},"/owner/audit/log":{"get":{"tags":["owner.audit"],"operationId":"getAuditLogEntries","parameters":[{"name":"begin","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"end","in":"query","required":true,"schema":{"type":"string","format":"date-time"}},{"name":"eventType","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OwnerAuditLogEntryDTO"}}}}}}}},"/owner/audit/log/{id}":{"get":{"tags":["owner.audit"],"operationId":"getAuditLogEntryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OwnerAuditLogEntryDTO"}}}}}}},"/owner/audit/event-type":{"get":{"tags":["owner.audit"],"operationId":"getEventTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string"}}}}}}}},"/consumer/data":{"get":{"tags":["consumer.data"],"operationId":"getDataItems_2","parameters":[{"name":"dataItemType","in":"query","required":true,"schema":{"type":"string"}},{"name":"ownerId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}}},"/consumer/data/{id}":{"get":{"tags":["consumer.data"],"operationId":"getDataItemById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DataItemDTO"}}}}}}},"/consumer/data/{id}/meta":{"get":{"tags":["consumer.data"],"operationId":"getMetaDataById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MetaDataDTO"}}}}}}},"/consumer/consent":{"get":{"tags":["consumer.consent"],"operationId":"getAllConsents_2","parameters":[{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["ACTIVE","REVOKED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ConsumerConsentDTO"}}}}}}}},"/consumer/consent/{id}":{"get":{"tags":["consumer.consent"],"operationId":"getConsentById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ConsumerConsentDTO"}}}}}}},"/catalog/data-item/type":{"get":{"tags":["catalog.dataitem"],"operationId":"getDataItemTypes","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DataItemTypeDTO"}}}}}}}},"/catalog/data-item/type/{id}/schema":{"get":{"tags":["catalog.dataitem"],"operationId":"getSchemaForDataItemType","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string","format":"binary"}}}}}}},"/owner/account/{id}":{"delete":{"tags":["owner"],"operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"}}}}},"components":{"schemas":{"UpdateDiseaseWarningSettingsDTO":{"type":"object","properties":{"isDiseaseReportProcessingEnabled":{"type":"boolean"},"isDiseaseWarningCreationEnabled":{"type":"boolean"},"isEmailNotificationEnabled":{"type":"boolean"}}},"DiseaseWarningSettingsDTO":{"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"},"isDiseaseReportProcessingEnabled":{"type":"boolean"},"isDiseaseWarningCreationEnabled":{"type":"boolean"},"isEmailNotificationEnabled":{"type":"boolean"}}},"UpdateOwnerProfileDTO":{"required":["preferredLanguage"],"type":"object","properties":{"preferredLanguage":{"type":"string"}}},"IdpProfileDTO":{"type":"object","properties":{"userId":{"type":"string"},"email":{"type":"string"},"firstname":{"type":"string"},"lastname":{"type":"string"}}},"OwnerProfileDTO":{"type":"object","properties":{"id":{"type":"string"},"idpProfile":{"$ref":"#/components/schemas/IdpProfileDTO"},"preferredLanguage":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"CreateDiseaseWarningSettingsDTO":{"type":"object","properties":{"isDiseaseReportProcessingEnabled":{"type":"boolean"},"isDiseaseWarningCreationEnabled":{"type":"boolean"},"isEmailNotificationEnabled":{"type":"boolean"}}},"CreateDataItemDTO":{"required":["metaData"],"type":"object","properties":{"metaData":{"$ref":"#/components/schemas/CreateMetaDataDTO"},"data":{"type":"object","additionalProperties":{"type":"object"}}}},"CreateMetaDataDTO":{"required":["dataItemType","ownerId"],"type":"object","properties":{"dataItemType":{"type":"string"},"ownerId":{"type":"string"}}},"DataItemDTO":{"required":["id","metaData"],"type":"object","properties":{"id":{"type":"string"},"metaData":{"$ref":"#/components/schemas/MetaDataDTO"},"data":{"type":"object","additionalProperties":{"type":"object"}}}},"MetaDataDTO":{"required":["createdAt","dataItemType","ownerId"],"type":"object","properties":{"dataItemType":{"type":"string"},"ownerId":{"type":"string"},"providerId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"ProviderCreateConsentRequestDTO":{"required":["dataItemTypes","ownerId"],"type":"object","properties":{"ownerId":{"type":"string"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"ProviderConsentRequestDTO":{"required":["dataItemTypes","id","ownerId","status"],"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}},"ProsumerCreateConsentRequestDTO":{"required":["consumedDataItemTypes","ownerId","providedDataItemTypes"],"type":"object","properties":{"ownerId":{"type":"string"},"consumedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"providedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]}}},"ProsumerConsentRequestDTO":{"required":["consumedDataItemTypes","id","ownerId","providedDataItemTypes","status"],"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"},"consumedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"providedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}},"CreateOwnerProfileDTO":{"required":["preferredLanguage"],"type":"object","properties":{"preferredLanguage":{"type":"string"}}},"ConsumerCreateConsentRequestDTO":{"required":["dataItemTypes","dataUsageStatement","ownerId","purpose"],"type":"object","properties":{"ownerId":{"type":"string"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]}}},"ConsumerConsentRequestDTO":{"required":["dataItemTypes","id","ownerId","status"],"type":"object","properties":{"id":{"type":"string"},"ownerId":{"type":"string"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}},"UpdateConsentRequestDTO":{"required":["id","status"],"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}},"UpdateConsentDTO":{"required":["id","status"],"type":"object","properties":{"id":{"type":"string"},"status":{"type":"string","enum":["ACTIVE","REVOKED"]}}},"ProviderConsentDTO":{"type":"object","properties":{"id":{"type":"string"},"consentRequestId":{"type":"string"},"ownerId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"status":{"type":"string","enum":["ACTIVE","REVOKED"]}}},"ProsumerConsentDTO":{"required":["consentRequestId","consumedDataItemTypes","createdAt","id","providedDataItemTypes","status"],"type":"object","properties":{"id":{"type":"string"},"consentRequestId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"consumedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"providedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["ACTIVE","REVOKED"]}}},"ConsentUserDTO":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}}},"OwnerConsentDTO":{"required":["consentRequestId","consumedDataItemTypes","createdAt","id","providedDataItemTypes","status"],"type":"object","properties":{"id":{"type":"string"},"consentRequestId":{"type":"string"},"actor":{"$ref":"#/components/schemas/ConsentUserDTO"},"createdAt":{"type":"string","format":"date-time"},"consumedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"providedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["ACTIVE","REVOKED"]}}},"OwnerConsentRequestDTO":{"required":["consumedDataItemTypes","createdAt","id","providedDataItemTypes","status"],"type":"object","properties":{"id":{"type":"string"},"requester":{"$ref":"#/components/schemas/ConsentUserDTO"},"createdAt":{"type":"string","format":"date-time"},"consumedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"providedDataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["PENDING","ACCEPTED","REJECTED","RETRACTED"]}}},"OwnerAuditLogEntryDTO":{"required":["actorId","dataItemTypes","entityId","eventType","id","operation","ownerId","timestamp"],"type":"object","properties":{"id":{"type":"string"},"timestamp":{"type":"string","format":"date-time"},"eventType":{"type":"string"},"ownerId":{"type":"string"},"entityId":{"type":"string"},"actorId":{"type":"string"},"operation":{"type":"string"},"dataItemTypes":{"type":"array","items":{"type":"string"}}}},"ConsumerConsentDTO":{"type":"object","properties":{"id":{"type":"string"},"consentRequestId":{"type":"string"},"ownerId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"dataItemTypes":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"dataUsageStatement":{"type":"string"},"purpose":{"type":"string","enum":["RESEARCH","SERVICE","OTHER"]},"status":{"type":"string","enum":["ACTIVE","REVOKED"]}}},"DataItemTypeDTO":{"required":["id","name"],"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"}}}},"securitySchemes":{"security_auth":{"type":"oauth2","flows":{"clientCredentials":{"tokenUrl":"http://localhost:8080/realms/master/protocol/openid-connect/token","scopes":{"consumer.consent:manage":"","consumer.data:read":"","owner.audit:manage":"","owner.consent:manage":"","owner.data:manage":"","owner.profile:manage":"","prosumer.consent:manage":"","prosumer.data:manage":"","provider.consent:manage":"","provider.data:write":""}},"authorizationCode":{"authorizationUrl":"http://localhost:8080/realms/master/protocol/openid-connect/auth","tokenUrl":"http://localhost:8080/realms/master/protocol/openid-connect/token","scopes":{"consumer.consent:manage":"","consumer.data:read":"","owner.audit:manage":"","owner.consent:manage":"","owner.data:manage":"","owner.profile:manage":"","prosumer.consent:manage":"","prosumer.data:manage":"","provider.consent:manage":"","provider.data:write":""}}}}}}}